Release 2.0.0

+++ lib/redmine_chat_telegram/commands/edit_issue_command.rb
@@ -207,14 +207,19 @@ module RedmineChatTelegram
 
       def change_issue_chat_name(name)
         if issue.telegram_group.present? && issue.telegram_group.telegram_id.present?
-          if account.user.allowed_to?(:edit_issues, issue.project)
-            chat_name = "chat##{issue.telegram_group.telegram_id.abs}"
-            cmd = "rename_chat #{chat_name} #{name}"
-            RedmineChatTelegram.socket_cli_command(cmd, logger)
-            executing_command.destroy
-            send_message(locale('chat_name_changed'))
+          if name.present?
+            if account.user.allowed_to?(:edit_issues, issue.project)
+              RedmineChatTelegram.run_cli_command('RenameChat', args: [issue.telegram_group.telegram_id.abs, name])
+              executing_command.destroy
+              send_message(locale('chat_name_changed'))
+            else
+              send_message(I18n.t('redmine_chat_telegram.bot.access_denied'))
+            end
           else
-            send_message(I18n.t('redmine_chat_telegram.bot.access_denied'))
+            result = RedmineChatTelegram.run_cli_command('BaseInfoChat', args: [issue.telegram_group.telegram_id.abs])
+            chat_info = JSON.parse(result)
+            chat_title = chat_info['chats']&.first['title']
+            send_message(chat_title.to_s)
           end
         else
           send_message(locale('chat_for_issue_not_exist'))
